requirements.txt
Django>3.2,<3.3
djangorestframework
django-taggit
Pillow
django-summernote
django-phonenumber-field
phonenumberslite
django-jalali-date
asgiref
pytz
sqlparse
bleach
six
webencodings